chore(angular): pnpm
