+++ packages/angular-schematics/ng-package.json
