+++ packages/angular-schematics/src/add/files/root/theme/variables.css
