+++ packages/angular-schematics/src/add/schema.json
