+++ packages/angular-schematics/src/collection.json
