+++ packages/angular-schematics/src/utils/package.ts
@@ -6,16 +6,18 @@ import { Tree } from '@angular-devkit/schematics';
 export function addPackageToPackageJson(host: Tree, type: string, pkg: string, version: string): Tree {
   if (host.exists('package.json')) {
     const sourceText = host.read('package.json')?.toString('utf-8');
-    const json = JSON.parse(sourceText);
-    if (!json[type]) {
-      json[type] = {};
-    }
+    if (sourceText) {
+      const json = JSON.parse(sourceText);
+      if (!json[type]) {
+        json[type] = {};
+      }
 
-    if (!json[type][pkg]) {
-      json[type][pkg] = version;
-    }
+      if (!json[type][pkg]) {
+        json[type][pkg] = version;
+      }
 
-    host.overwrite('package.json', JSON.stringify(json, null, 2));
+      host.overwrite('package.json', JSON.stringify(json, null, 2));
+    }
   }
 
   return host;
