+++ packages/angular/standalone/ng-package.json
@@ -1,5 +1,5 @@
 {
   "lib": {
     "entryFile": "src/index.ts"
-  },
+  }
 }
