+++ pnpm-workspace.yaml
@@ -1,9 +1,10 @@
 packages:
-  - 'core'
-  - 'docs'
-  - 'packages/angular'
-  - 'packages/angular-server'
-  - 'packages/vue'
-  - 'packages/vue-router'
-  - 'packages/react'
-  - 'packages/react-router'
+  - "core"
+  - "docs"
+  - "packages/angular"
+  - "packages/angular-server"
+  - "packages/angular-schematics"
+  - "packages/vue"
+  - "packages/vue-router"
+  - "packages/react"
+  - "packages/react-router"
